+++ changelog.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `install-bx-modules` so you can do multi-list installations
+
 ## [1.1.0] - 2024-07-08
 
 ### Added